+++ client/mdns_discovery.go
@@ -55,14 +55,6 @@ func NewMDNSDiscovery(service string, timeout time.Duration, watchInterval time.
 	return d, nil
 }
 
-// NewMDNSDiscoveryTemplate returns a new MDNSDiscovery template.
-func NewMDNSDiscoveryTemplate(timeout time.Duration, watchInterval time.Duration, domain string) (ServiceDiscovery, error) {
-	if domain == "" {
-		domain = "local."
-	}
-	return &MDNSDiscovery{Timeout: timeout, WatchInterval: watchInterval, domain: domain}, nil
-}
-
 // Clone clones this ServiceDiscovery with new servicePath.
 func (d *MDNSDiscovery) Clone(servicePath string) (ServiceDiscovery, error) {
 	return NewMDNSDiscovery(servicePath, d.Timeout, d.WatchInterval, d.domain)